+++ Tests/EventListener/WebDebugToolbarListenerTest.php
@@ -16,8 +16,10 @@ use PHPUnit\Framework\Attributes\Depends;
 use PHPUnit\Framework\TestCase;
 use Symfony\Bundle\WebProfilerBundle\Csp\ContentSecurityPolicyHandler;
 use Symfony\Bundle\WebProfilerBundle\EventListener\WebDebugToolbarListener;
+use Symfony\Component\HttpFoundation\EventStreamResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\ServerEvent;
 use Symfony\Component\HttpKernel\DataCollector\DumpDataCollector;
 use Symfony\Component\HttpKernel\Event\ResponseEvent;
 use Symfony\Component\HttpKernel\HttpKernelInterface;
@@ -399,6 +401,88 @@ class WebDebugToolbarListenerTest extends TestCase
         $listener->onKernelResponse($event);
 
         $this->assertSame('0', $response->headers->get('Symfony-Debug-Toolbar-Replace'));
+    }
+
+    public function testEventStreamResponseHasDebugEvents()
+    {
+        if (!class_exists(EventStreamResponse::class)) {
+            self::markTestSkipped('This test requires symfony/http-foundation >= 7.3');
+        }
+
+        $request = new Request();
+        $response = new EventStreamResponse(
+            fn () => yield new ServerEvent('some data'),
+            headers: [
+                'X-Debug-Token' => 'aabbcc',
+                'X-Debug-Token-Link' => 'test://foobar',
+            ],
+        );
+        $event = new ResponseEvent($this->createMock(Kernel::class), $request, HttpKernelInterface::MAIN_REQUEST, $response);
+
+        $listener = new WebDebugToolbarListener($this->getTwigMock());
+
+        $listener->onKernelResponse($event);
+
+        $this->expectOutputString(
+            <<<'EVENTSTREAM'
+                event: symfony:debug:started
+                data: aabbcc
+                data: test://foobar
+
+                data: some data
+
+                event: symfony:debug:finished
+                data: -
+
+
+                EVENTSTREAM
+        );
+        $response->send(false);
+    }
+
+    public function testEventStreamResponseHasDebugEventForException()
+    {
+        if (!class_exists(EventStreamResponse::class)) {
+            self::markTestSkipped('This test requires symfony/http-foundation >= 7.3');
+        }
+
+        $request = new Request();
+        $response = new EventStreamResponse(
+            function () {
+                yield new ServerEvent('some data');
+                throw new \RuntimeException('Something went wrong');
+            },
+            headers: [
+                'X-Debug-Token' => 'aabbcc',
+                'X-Debug-Token-Link' => 'test://foobar',
+            ],
+        );
+        $event = new ResponseEvent($this->createMock(Kernel::class), $request, HttpKernelInterface::MAIN_REQUEST, $response);
+
+        $listener = new WebDebugToolbarListener($this->getTwigMock());
+
+        $listener->onKernelResponse($event);
+
+        $this->expectOutputString(
+            <<<'EVENTSTREAM'
+                event: symfony:debug:started
+                data: aabbcc
+                data: test://foobar
+
+                data: some data
+
+                event: symfony:debug:error
+                data: error
+
+                event: symfony:debug:finished
+                data: -
+
+
+                EVENTSTREAM
+        );
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('Something went wrong');
+        $response->send(false);
     }
 
     protected function getTwigMock($render = 'WDT')